put everything in one pipeline
