+++ clinicadl/generate/generate.py
@@ -659,374 +659,6 @@ def generate_hypometabolic_dataset(
     )
 
 
-def generate_motion_dataset(
-    caps_directory: Path,
-    output_dir: Path,
-    n_proc: int,
-    tsv_path: Optional[str] = None,
-    preprocessing: str = "t1-linear",
-    multi_cohort: bool = False,
-    uncropped_image: bool = False,
-    tracer: str = "fdg",
-    suvr_reference_region: str = "pons",
-    translation: List = [2, 4],
-    rotation: List = [2, 4],
-    num_transforms: int = 2,
-):
-    """
-    Generates a dataset, based on the images of the CAPS directory, where
-    all the images are corrupted with motion artefacts using the image-based simulation of torchio.
-    Args:
-        caps_directory:  Path
-            Path to the CAPS directory.
-        output_dir: Path
-            Folder containing the synthetic dataset in CAPS format.
-        n_proc: int
-            Number of cores used during the task.
-        tsv_path: Path
-            Path to tsv file of list of subjects/sessions.
-        preprocessing: str
-            Preprocessing performed. Must be in ['linear', 'extensive'].
-        multi_cohort: bool
-            If True caps_directory is the path to a TSV file linking cohort names and paths.
-        uncropped_image: bool
-            If True the uncropped image of `t1-linear` or `pet-linear` will be used.
-        tracer: str
-            Name of the tracer when using `pet-linear` preprocessing.
-        suvr_reference_region: str
-            Name of the reference region when using `pet-linear` preprocessing.
-        translation: List
-            Translation range in mm of simulated movements.
-        rotation : List
-            Rotation range in degree of simulated movement.
-        num_transformes: int
-            Number of simulated movements.
-    Returns:
-        Folder structure where images are stored in CAPS format.
-    """
-
-    commandline_to_json(
-        {
-            "output_dir": output_dir,
-            "caps_dir": caps_directory,
-            "preprocessing": preprocessing,
-        }
-    )
-
-    # Transform caps_directory in dict
-    caps_dict = CapsDataset.create_caps_dict(caps_directory, multi_cohort=multi_cohort)
-    # Read DataFrame
-    data_df = load_and_check_tsv(tsv_path, caps_dict, output_dir)
-    # Create subjects dir
-    (output_dir / "subjects").mkdir(parents=True, exist_ok=True)
-
-    # Output tsv file
-    columns = ["participant_id", "session_id", "diagnosis"]
-    output_df = pd.DataFrame(columns=columns)
-
-    # Find appropriate preprocessing file type
-    file_type = find_file_type(
-        preprocessing, uncropped_image, tracer, suvr_reference_region
-    )
-
-    def create_motion_image(data_idx, output_df):
-        participant_id = data_df.loc[data_idx, "participant_id"]
-        session_id = data_df.loc[data_idx, "session_id"]
-        cohort = data_df.loc[data_idx, "cohort"]
-        image_path = Path(
-            clinica_file_reader(
-                [participant_id], [session_id], caps_dict[cohort], file_type
-            )[0][0]
-        )
-        input_filename = image_path.name
-        filename_pattern = "_".join(input_filename.split("_")[2::])
-
-        subject_id = participant_id.split("-")[1]
-
-        motion_image_nii_dir = (
-            output_dir
-            / "subjects"
-            / f"sub-MOTION{subject_id}"
-            / session_id
-            / preprocessing
-        )
-        motion_image_nii_filename = (
-            f"sub-MOTION{subject_id}_{session_id}_{filename_pattern}"
-        )
-
-        motion_image_nii_dir.mkdir(parents=True, exist_ok=True)
-
-        motion = tio.RandomMotion(
-            degrees=(rotation[0], rotation[1]),
-            translation=(translation[0], translation[1]),
-            num_transforms=num_transforms,
-        )
-
-        motion_image = motion(tio.ScalarImage(image_path))
-        motion_image.save(motion_image_nii_dir / motion_image_nii_filename)
-
-        # Append row to output tsv
-        row = [f"sub-MOTION{subject_id}", session_id, "motion"]
-        row_df = pd.DataFrame([row], columns=columns)
-        output_df = pd.concat([output_df, row_df])
-
-        return output_df
-
-    results_df = Parallel(n_jobs=n_proc)(
-        delayed(create_motion_image)(data_idx, output_df)
-        for data_idx in range(len(data_df))
-    )
-    output_df = pd.DataFrame()
-    for result in results_df:
-        output_df = pd.concat([result, output_df])
-
-    output_df.to_csv(output_dir / "data.tsv", sep="\t", index=False)
-
-    write_missing_mods(output_dir, output_df)
-
-    logger.info(
-        f"Images corrupted with motion artefacts were generated at {output_dir}"
-    )
-
-
-def generate_contrast_dataset(
-    caps_directory: Path,
-    output_dir: Path,
-    n_proc: int,
-    tsv_path: Optional[str] = None,
-    preprocessing: str = "t1-linear",
-    multi_cohort: bool = False,
-    uncropped_image: bool = False,
-    tracer: str = "fdg",
-    suvr_reference_region: str = "pons",
-    gamma: List = [-0.2, -0.05],
-):
-    """
-    Generates a dataset, based on the images of the CAPS directory, where
-    all the images are corrupted with contrast artefacts using the gamma simulation of torchio.
-    Args:
-        caps_directory: Path
-            Path to the CAPS directory.
-        output_dir: Path
-            Folder containing the synthetic dataset in CAPS format.
-        n_proc: int
-            Number of cores used during the task.
-        tsv_path: Path
-            Path to tsv file of list of subjects/sessions.
-        preprocessing: str
-            Preprocessing performed. Must be in ['linear', 'extensive'].
-        multi_cohort: bool
-            If True caps_directory is the path to a TSV file linking cohort names and paths.
-        uncropped_image: bool
-            If True the uncropped image of `t1-linear` or `pet-linear` will be used.
-        tracer: str
-            Name of the tracer when using `pet-linear` preprocessing.
-        suvr_reference_region: str
-            Name of the reference region when using `pet-linear` preprocessing.
-        gamma: List
-            Range to compute gamma exponentiation.
-    Returns:
-        Folder structure where images are stored in CAPS format.
-    """
-
-    commandline_to_json(
-        {
-            "output_dir": output_dir,
-            "caps_dir": caps_directory,
-            "preprocessing": preprocessing,
-        }
-    )
-
-    # Transform caps_directory in dict
-    caps_dict = CapsDataset.create_caps_dict(caps_directory, multi_cohort=multi_cohort)
-    # Read DataFrame
-    data_df = load_and_check_tsv(tsv_path, caps_dict, output_dir)
-    # Create subjects dir
-    (output_dir / "subjects").mkdir(parents=True, exist_ok=True)
-
-    # Output tsv file
-    columns = ["participant_id", "session_id", "diagnosis"]
-    output_df = pd.DataFrame(columns=columns)
-
-    # Find appropriate preprocessing file type
-    file_type = find_file_type(
-        preprocessing, uncropped_image, tracer, suvr_reference_region
-    )
-
-    def create_contrast_image(data_idx, output_df):
-        participant_id = data_df.loc[data_idx, "participant_id"]
-        session_id = data_df.loc[data_idx, "session_id"]
-        cohort = data_df.loc[data_idx, "cohort"]
-
-        image_path = Path(
-            clinica_file_reader(
-                [participant_id], [session_id], caps_dict[cohort], file_type
-            )[0][0]
-        )
-        input_filename = image_path.name
-        filename_pattern = "_".join(input_filename.split("_")[2::])
-
-        subject_id = participant_id.split("-")[1]
-
-        contrast_image_nii_dir = (
-            output_dir
-            / "subjects"
-            / f"sub-CONT{subject_id}"
-            / session_id
-            / preprocessing
-        )
-        contrast_image_nii_filename = (
-            f"sub-CONT{subject_id}_{session_id}_{filename_pattern}"
-        )
-
-        contrast_image_nii_dir.mkdir(parents=True, exist_ok=True)
-
-        contrast = tio.RandomGamma(log_gamma=(gamma[0], gamma[1]))
-
-        contrast_image = contrast(tio.ScalarImage(image_path))
-        contrast_image.save(contrast_image_nii_dir / contrast_image_nii_filename)
-
-        # Append row to output tsv
-        row = [f"sub-CONT{subject_id}", session_id, "contrast"]
-        row_df = pd.DataFrame([row], columns=columns)
-        output_df = pd.concat([output_df, row_df])
-
-        return output_df
-
-    results_df = Parallel(n_jobs=n_proc)(
-        delayed(create_contrast_image)(data_idx, output_df)
-        for data_idx in range(len(data_df))
-    )
-    output_df = pd.DataFrame()
-    for result in results_df:
-        output_df = pd.concat([result, output_df])
-
-    output_df.to_csv(output_dir / "data.tsv", sep="\t", index=False)
-
-    write_missing_mods(output_dir, output_df)
-
-    logger.info(
-        f"Images corrupted with contrast artefacts were generated at {output_dir}"
-    )
-
-
-def generate_noise_dataset(
-    caps_directory: Path,
-    output_dir: Path,
-    n_proc: int,
-    tsv_path: Optional[str] = None,
-    preprocessing: str = "t1-linear",
-    multi_cohort: bool = False,
-    uncropped_image: bool = False,
-    tracer: str = "fdg",
-    suvr_reference_region: str = "pons",
-    noise_std: List = [5, 15],
-):
-    """
-    Generates a dataset, based on the images of the CAPS directory, where
-    all the images are corrupted with noise artefacts using the gaussian noise simulation of torchio.
-    Args:
-        caps_directory: Path
-            Path to the CAPS directory.
-        output_dir: Path
-            Folder containing the synthetic dataset in CAPS format.
-        n_proc: int
-            Number of cores used during the task.
-        tsv_path: Path
-            Path to tsv file of list of subjects/sessions.
-        preprocessing: str
-            Preprocessing performed. Must be in ['linear', 'extensive'].
-        multi_cohort: bool
-            If True caps_directory is the path to a TSV file linking cohort names and paths.
-        uncropped_image: bool
-            If True the uncropped image of `t1-linear` or `pet-linear` will be used.
-        tracer: str
-            Name of the tracer when using `pet-linear` preprocessing.
-        suvr_reference_region: str
-            Name of the reference region when using `pet-linear` preprocessing.
-    Returns:
-        Folder structure where images are stored in CAPS format.
-    """
-
-    commandline_to_json(
-        {
-            "output_dir": output_dir,
-            "caps_dir": caps_directory,
-            "preprocessing": preprocessing,
-        }
-    )
-
-    # Transform caps_directory in dict
-    caps_dict = CapsDataset.create_caps_dict(caps_directory, multi_cohort=multi_cohort)
-    # Read DataFrame
-    data_df = load_and_check_tsv(tsv_path, caps_dict, output_dir)
-    # Create subjects dir
-    (output_dir / "subjects").mkdir(parents=True, exist_ok=True)
-
-    # Output tsv file
-    columns = ["participant_id", "session_id", "diagnosis"]
-    output_df = pd.DataFrame(columns=columns)
-
-    # Find appropriate preprocessing file type
-    file_type = find_file_type(
-        preprocessing, uncropped_image, tracer, suvr_reference_region
-    )
-
-    def create_noise_image(data_idx, output_df):
-        participant_id = data_df.loc[data_idx, "participant_id"]
-        session_id = data_df.loc[data_idx, "session_id"]
-        cohort = data_df.loc[data_idx, "cohort"]
-
-        image_path = Path(
-            clinica_file_reader(
-                [participant_id], [session_id], caps_dict[cohort], file_type
-            )[0][0]
-        )
-        input_filename = image_path.name
-        filename_pattern = "_".join(input_filename.split("_")[2::])
-
-        subject_id = participant_id.split("-")[1]
-
-        noise_image_nii_dir = (
-            output_dir
-            / "subjects"
-            / f"sub-NOIS{subject_id}"
-            / session_id
-            / preprocessing
-        )
-        noise_image_nii_filename = (
-            f"sub-NOIS{subject_id}_{session_id}_{filename_pattern}"
-        )
-
-        noise_image_nii_dir.mkdir(parents=True, exist_ok=True)
-
-        noise = tio.RandomNoise(std=(noise_std[0], noise_std[1]))
-
-        noise_image = noise(tio.ScalarImage(image_path))
-        noise_image.save(noise_image_nii_dir / noise_image_nii_filename)
-
-        # Append row to output tsv
-        row = [f"sub-NOIS{subject_id}", session_id, "noise"]
-        row_df = pd.DataFrame([row], columns=columns)
-        output_df = pd.concat([output_df, row_df])
-
-        return output_df
-
-    results_df = Parallel(n_jobs=n_proc)(
-        delayed(create_noise_image)(data_idx, output_df)
-        for data_idx in range(len(data_df))
-    )
-    output_df = pd.DataFrame()
-    for result in results_df:
-        output_df = pd.concat([result, output_df])
-
-    output_df.to_csv(output_dir / "data.tsv", sep="\t", index=False)
-
-    write_missing_mods(output_dir, output_df)
-
-    logger.info(f"Images corrupted with noise artefacts were generated at {output_dir}")
-
-
 def generate_artifacts_dataset(
     caps_directory: Path,
     output_dir: Path,
@@ -1037,11 +669,14 @@ def generate_artifacts_dataset(
     uncropped_image: bool = False,
     tracer: str = "fdg",
     suvr_reference_region: str = "pons",
-    #    translation: List = [2, 4],
-    #    rotation: List = [2, 4],
-    #    num_transforms: int = 2,
+    contrast: bool = False,
     gamma: List = [-0.2, -0.05],
-    #    noise_std: List = [5, 15]
+    motion: bool = False,
+    translation: List = [2, 4],
+    rotation: List = [2, 4],
+    num_transforms: int = 2,
+    noise: bool = False,
+    noise_std: List = [5, 15],
 ):
     """
     Generates a dataset, based on the images of the CAPS directory, where
@@ -1103,6 +738,13 @@ def generate_artifacts_dataset(
     file_type = find_file_type(
         preprocessing, uncropped_image, tracer, suvr_reference_region
     )
+    artifacts_list = []
+    if motion:
+        artifacts_list.append("motion")
+    if contrast:
+        artifacts_list.append("contrast")
+    if noise:
+        artifacts_list.append("noise")
 
     def create_artifacts_image(data_idx, output_df):
         participant_id = data_df.loc[data_idx, "participant_id"]
@@ -1115,46 +757,53 @@ def generate_artifacts_dataset(
         )
         input_filename = image_path.name
         filename_pattern = "_".join(input_filename.split("_")[2::])
-
-        subject_id = participant_id.split("-")[1]
+        subject_name = input_filename.split("_")[:1][0]
+        session_name = input_filename.split("_")[1:2][0]
 
         artif_image_nii_dir = (
-            output_dir
-            / "subjects"
-            / f"sub-ARTIF{subject_id}"
-            / session_id
-            / preprocessing
+            output_dir / "subjects" / subject_name / session_name / preprocessing
         )
-        artif_image_nii_filename = (
-            f"sub-ARTIF{subject_id}_{session_id}_{filename_pattern}"
-        )
-
         artif_image_nii_dir.mkdir(parents=True, exist_ok=True)
 
-        artifacts = tio.RandomGamma(log_gamma=(gamma[0], gamma[1]))
+        artifacts_tio = []
+        arti_ext = ""
+        for artif in artifacts_list:
+            if artif == "motion":
+                artifacts_tio.append(
+                    tio.RandomMotion(
+                        degrees=(rotation[0], rotation[1]),
+                        translation=(translation[0], translation[1]),
+                        num_transforms=num_transforms,
+                    )
+                )
+                arti_ext += "mot-"
+            elif artif == "noise":
+                artifacts_tio.append(
+                    tio.RandomNoise(
+                        std=(noise_std[0], noise_std[1]),
+                    )
+                )
+                arti_ext += "noi-"
+            elif artif == "contrast":
+                artifacts_tio.append(tio.RandomGamma(log_gamma=(gamma[0], gamma[1])))
+                arti_ext += "con-"
 
-        """
-        artifacts = tio.transforms.Compose(
-            [
-                tio.RandomGamma(
-                    log_gamma=(gamma[0], gamma[1])
-                ),
-                tio.RandomMotion(
-                    degrees=(rotation[0], rotation[1]),
-                    translation=(translation[0], translation[1]),
-                    num_transforms=num_transforms,
-                ),
-                tio.RandomNoise(std=(noise_std[0], noise_std[1]),
-                ),
-            ]
-        )
-        """
+        if filename_pattern.endswith(".nii.gz"):
+            file_suffix = ".nii.gz"
+            filename_pattern = Path(Path(filename_pattern).stem).stem
+        elif filename_pattern.endswith(".nii"):
+            file_suffix = ".nii"
+            filename_pattern = Path(filename_pattern).stem
+
+        artif_image_nii_filename = f"{subject_name}_{session_name}_{filename_pattern}_art-{arti_ext[:-1]}{file_suffix}"
+
+        artifacts = tio.transforms.Compose(artifacts_tio)
 
         artif_image = artifacts(tio.ScalarImage(image_path))
         artif_image.save(artif_image_nii_dir / artif_image_nii_filename)
 
         # Append row to output tsv
-        row = [f"sub-ARTIF{subject_id}", session_id, "motion"]
+        row = [subject_name, session_name, artifacts_list]
         row_df = pd.DataFrame([row], columns=columns)
         output_df = pd.concat([output_df, row_df])
