+++ clinicadl/generate/generate_artifacts_cli.py
@@ -1,11 +1,10 @@
-from typing import Dict, List, Optional, Tuple, Union
-
 import click
 
+from clinicadl.generate.generate import generate_artifacts_dataset
 from clinicadl.utils import cli_param
 
 
-@click.command(name="trivial_artifacts", no_args_is_help=True)
+@click.command(name="artifacts", no_args_is_help=True)
 @cli_param.argument.caps_directory
 @cli_param.argument.generated_caps
 @cli_param.option.n_proc
@@ -14,9 +13,15 @@ from clinicadl.utils import cli_param
 @cli_param.option.use_uncropped_image
 @cli_param.option.tracer
 @cli_param.option.suvr_reference_region
-
 ##############
 # Contrast
+@click.option(
+    "--contrast/--no-contrast",
+    type=bool,
+    default=False,
+    is_flag=True,
+    help="",
+)
 @click.option(
     "--gamma",
     type=float,
@@ -24,39 +29,49 @@ from clinicadl.utils import cli_param
     default=[-0.2, -0.05],
     help="Range between -1 and 1 for gamma augmentation",
 )
-
-
 # Motion
-# @click.option(
-#   "--translation",
-#    type=float,
-#    multiple=2,
-#    default=[2, 4],
-#    help="Range in mm for the translation",
-# )
-# @click.option(
-#    "--rotation",
-##    type=float,
-#    multiple=2,
-#    default=[2, 4],
-#    help="Range in degree for the rotation",
-# )
-# @click.option(
-#    "--num_transforms",
-#    type=int,
-#    default=2,
-#    help="Number of transforms",
-# )
-##Noise
-# @click.option(
-#    "--noise_std",
-#    type=float,
-#    multiple=2,
-#    default=[5, 15],
-#    help="Range for noise standard deviation",
-# )
-
-
+@click.option(
+    "--motion/--no-motion",
+    type=bool,
+    default=False,
+    is_flag=True,
+    help="",
+)
+@click.option(
+    "--translation",
+    type=float,
+    multiple=2,
+    default=[2, 4],
+    help="Range in mm for the translation",
+)
+@click.option(
+    "--rotation",
+    #    type=float,
+    multiple=2,
+    default=[2, 4],
+    help="Range in degree for the rotation",
+)
+@click.option(
+    "--num_transforms",
+    type=int,
+    default=2,
+    help="Number of transforms",
+)
+# Noise
+@click.option(
+    "--noise/--no-noise",
+    type=bool,
+    default=False,
+    is_flag=True,
+    help="",
+)
+@click.option(
+    "--noise_std",
+    type=float,
+    multiple=2,
+    default=[5, 15],
+    help="Range for noise standard deviation",
+)
 def cli(
     caps_directory,
     generated_caps_directory,
@@ -65,18 +80,20 @@ def cli(
     use_uncropped_image,
     tracer,
     suvr_reference_region,
+    contrast,
     gamma,
-    #    translation,
-    #    rotation,
-    #    num_transforms,
-    #    noise_std,
+    motion,
+    translation,
+    rotation,
+    num_transforms,
+    noise,
+    noise_std,
     n_proc,
 ):
     """Generation of trivial dataset with addition of synthetic artifacts.
     CAPS_DIRECTORY is the CAPS folder from where input brain images will be loaded.
     GENERATED_CAPS_DIRECTORY is a CAPS folder where the trivial dataset will be saved.
     """
-    from .generate import generate_artifacts_dataset
 
     generate_artifacts_dataset(
         caps_directory=caps_directory,
@@ -86,11 +103,14 @@ def cli(
         uncropped_image=use_uncropped_image,
         tracer=tracer,
         suvr_reference_region=suvr_reference_region,
+        contrast=contrast,
         gamma=gamma,
-        #        translation=translation,
-        #        rotation=rotation,
-        #        num_trasnforms=num_transforms,
-        #        noise_std=noise_std
+        motion=motion,
+        translation=translation,
+        rotation=rotation,
+        num_transforms=num_transforms,
+        noise=noise,
+        noise_std=noise_std,
         n_proc=n_proc,
     )
