+++ clinicadl/generate/generate_cli.py
@@ -1,13 +1,11 @@
 import click
 
+from clinicadl.generate.generate_artifacts_cli import cli as generate_artifacts_cli
+
 from .generate_hypometabolic_cli import cli as generate_hypo_cli
 from .generate_random_cli import cli as generate_random_cli
 from .generate_shepplogan_cli import cli as generate_shepplogan_cli
 from .generate_trivial_cli import cli as generate_trivial_cli
-from .generate_trivial_contrast_cli import cli as generate_trivial_contrast_cli
-from .generate_trivial_motion_cli import cli as generate_trivial_motion_cli
-from .generate_trivial_noise_cli import cli as generate_trivial_noise_cli
-from .generate_trivial_artifacts_cli import cli as generate_trivial_artifacts_cli
 
 
 class RegistrationOrderGroup(click.Group):
@@ -27,10 +25,7 @@ cli.add_command(generate_random_cli)
 cli.add_command(generate_trivial_cli)
 cli.add_command(generate_shepplogan_cli)
 cli.add_command(generate_hypo_cli)
-cli.add_command(generate_trivial_motion_cli)
-cli.add_command(generate_trivial_contrast_cli)
-cli.add_command(generate_trivial_noise_cli)
-cli.add_command(generate_trivial_artifacts_cli)
+cli.add_command(generate_artifacts_cli)
 
 
 if __name__ == "__main__":
